backend/requirements.txt
fastapi
uvicorn[standard]
msal
azure-identity
python-dotenv
requests
openai
sqlalchemy
psycopg2-binary
redis
pytest
pytest-mock
pytest-asyncio
pytest-cov
httpx
sentence-transformers
scikit-learn
transformers
beautifulsoup4
html2text
demjson3